requirements.txt
tensorflow
numpy
# matplotlab
opencv-python
baidu-aip
